♻️ S96Y5J prompt: decompose registry factories
