+++ .agentplane/tasks/202605290522-S96Y5J/pr/diffstat.txt
@@ -1,0 +1,3 @@
+ .../src/runtime/prompt-modules/registry.factory.ts | 227 ++++++++++++++++++++
+ .../src/runtime/prompt-modules/registry.ts         | 233 +--------------------
+ 2 files changed, 234 insertions(+), 226 deletions(-)
