+++ .agentplane/tasks/202605290522-S96Y5J/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 S96Y5J task: Prompt module registry decomposition [202605290522-S96Y5J]
