+++ .agentplane/tasks/202605290522-S96Y5J/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290522-S96Y5J/prompt-module-registry-decomposition",
+  "created_at": "2026-05-29T05:22:36.686Z",
+  "diffstat_sha256": "sha256:b8139f2a08a62a871c2c0161ca2b3e715e9df4be93bddf90f6f4f401e510a3ee",
+  "last_verified_at": "2026-05-29T05:25:51.201Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290522-S96Y5J",
+  "updated_at": "2026-05-29T05:22:36.686Z",
+  "verify": {
+    "status": "pass"
+  }
+}
