+++ .agentplane/tasks/202605290522-S96Y5J/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T05:22:36.686Z
+
+## Task
+
+- Task: `202605290522-S96Y5J`
+- Title: Prompt module registry decomposition
+- Status: DOING
+- Branch: `task/202605290522-S96Y5J/prompt-module-registry-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290522-S96Y5J/README.md`
+
+## Verification
+
+- State: ok
+- Note: Prompt module registry construction helpers extracted into registry.factory.ts; registry.ts reduced to 216 lines while preserving framework prompt module graph exports.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T05:22:36.686Z
+- Branch: task/202605290522-S96Y5J/prompt-module-registry-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/runtime/prompt-modules/registry.factory.ts | 227 ++++++++++++++++++++
+ .../src/runtime/prompt-modules/registry.ts         | 233 +--------------------
+ 2 files changed, 234 insertions(+), 226 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
